+++ ui/main/game/galactic_war/cards/gwaio_upgrade_subcommander_fabber.js
@@ -18,7 +18,9 @@ define(["coui://ui/mods/com.pa.quitch.gwaioverhaul/shared/cards.js"], function (
     getContext: gwoCard.getContext,
     deal: function (system, context, inventory) {
       var chance = 0;
-      if (inventory.minions().length > 0) {
+      var galaxy = model.game().galaxy();
+      var ai = galaxy.stars()[galaxy.origin()].system().gwaio.ai;
+      if (inventory.minions().length > 0 && ai !== "Queller") {
         chance = 30;
       }
       return { chance: chance };