+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## Unreleased
+
+### Fixed
+
+- Don't deal Sub Commander Fabber Tech if the AI is Queller
+
 ## v5.34.0 - 2022-06-11
 
 ### Changed